chore: close v4.7.26 council gaps, add CHECKPOINTS.yaml, seal v4.7.40

+++ antigravity_harness/execution/adapter_base.py
@@ -89,7 +89,7 @@ class Fill:
     client_order_id: str
     symbol: str
     side: OrderSide
-    filled_qty: int
+    filled_qty: float
     fill_price: Decimal
     fill_time_utc: datetime
     commission_usd: float
